Split out common function in separate file
