+++ ansible/plugins/modules/vault_load_secrets.py
@@ -51,12 +51,10 @@ files.region2:
 """
 
 import os
-import subprocess
-import time
-from collections.abc import MutableMapping
 
 import yaml
 from ansible.module_utils.basic import AnsibleModule
+from ansible.module_utils.load_secrets_common import parse_values, get_version, run_command, flatten
 
 
 ANSIBLE_METADATA = {
@@ -122,106 +120,13 @@ EXAMPLES = """
 """
 
 
-def parse_values(values_file):
-    """
-    Parses a values-secrets.yaml file (usually placed in ~)
-    and returns a Python Obect with the parsed yaml.
-
-    Parameters:
-        values_file(str): The path of the values-secrets.yaml file
-        to be parsed.
-
-    Returns:
-        secrets_yaml(obj): The python object containing the parsed yaml
-    """
-    with open(values_file, "r", encoding="utf-8") as file:
-        secrets_yaml = yaml.safe_load(file.read())
-    if secrets_yaml is None:
-        return {}
-    return secrets_yaml
-
-
-def get_version(syaml):
-    """
-    Return the version: of the parsed yaml object. If it does not exist
-    return 1.0
-
-    Returns:
-        ret(str): The version value in of the top-level 'version:' key
-    """
-    return syaml.get("version", "1.0")
-
-
-def run_command(command, attempts=1, sleep=3):
-    """
-    Runs a command on the host ansible is running on. A failing command
-    will raise an exception in this function directly (due to check=True)
-
-    Parameters:
-        command(str): The command to be run.
-        attempts(int): Number of times to retry in case of Error (defaults to 1)
-        sleep(int): Number of seconds to wait in between retry attempts (defaults to 3s)
-
-    Returns:
-        ret(subprocess.CompletedProcess): The return value from run()
-    """
-    for attempt in range(attempts):
-        try:
-            ret = subprocess.run(
-                command,
-                shell=True,
-                env=os.environ.copy(),
-                stdout=subprocess.PIPE,
-                stderr=subprocess.PIPE,
-                universal_newlines=True,
-                check=True,
-            )
-            return ret
-        except subprocess.CalledProcessError as e:
-            # We reached maximum nr of retries. Re-raise the last error
-            if attempt >= attempts - 1:
-                raise e
-            time.sleep(sleep)
-
-
-def flatten(dictionary, parent_key=False, separator="."):
-    """
-    Turn a nested dictionary into a flattened dictionary and also
-    drop any key that has 'None' as their value
-
-    Parameters:
-        dictionary(dict): The dictionary to flatten
-
-        parent_key(str): The string to prepend to dictionary's keys
-
-        separator(str): The string used to separate flattened keys
-
-    Returns:
-
-        dictionary: A flattened dictionary where the keys represent the
-        path to reach the leaves
-    """
-
-    items = []
-    for key, value in dictionary.items():
-        new_key = str(parent_key) + separator + key if parent_key else key
-        if isinstance(value, MutableMapping):
-            items.extend(flatten(value, new_key, separator).items())
-        elif isinstance(value, list):
-            for k, v in enumerate(value):
-                items.extend(flatten({str(k): v}, new_key).items())
-        else:
-            if value is not None:
-                items.append((new_key, value))
-    return dict(items)
-
 
 
 # NOTE(bandini): we shell out to oc exec it because of
 # https://github.com/ansible-collections/kubernetes.core/issues/506 and
 # https://github.com/kubernetes/kubernetes/issues/89899. Until those are solved
 # it makes little sense to invoke the APIs via the python wrappers
-def inject_secrets(module, syaml, namespace, pod, basepath):
+def inject_secrets(module, syaml, namespace, pod, basepath, get_secrets_vault_paths_func):
     """
     Walks a secrets yaml object and injects all the secrets into the vault via 'oc exec' calls
 
@@ -241,7 +146,7 @@ def inject_secrets(module, syaml, namespace, pod, basepath):
         counter(int): The number of secrets injected
     """
     counter = 0
-    for i in get_secrets_vault_paths(module, syaml, "secrets"):
+    for i in get_secrets_vault_paths_func(module, syaml, "secrets"):
         path = f"{basepath}/{i[1]}"
         for secret in syaml[i[0]] or []:
             properties = ""
@@ -255,7 +160,7 @@ def inject_secrets(module, syaml, namespace, pod, basepath):
             run_command(cmd, attempts=3)
             counter += 1
 
-    for i in get_secrets_vault_paths(module, syaml, "files"):
+    for i in get_secrets_vault_paths_func(module, syaml, "files"):
         path = f"{basepath}/{i[1]}"
         for filekey in syaml[i[0]] or []:
             file = os.path.expanduser(syaml[i[0]][filekey])
@@ -332,7 +237,7 @@ def run(module):
     if check_missing_secrets:
         check_for_missing_secrets(module, syaml, values_secret_template)
 
-    nr_secrets = inject_secrets(module, secrets, namespace, pod, basepath)
+    nr_secrets = inject_secrets(module, secrets, namespace, pod, basepath, get_secrets_vault_paths)
     results["failed"] = False
     results["changed"] = True
     results["msg"] = f"{nr_secrets} secrets injected"
